--- lametro/templates/captcha/includes/js_v2_checkbox.html
@@ -1,9 +1,0 @@
-{# The provided implementation caters for only one reCAPTCHA on a page. Override this template and its logic as needed. #}
-<script src="https://{{ recaptcha_domain }}/recaptcha/api.js{% if api_params %}?{{ api_params }}{% endif %}"></script>
-<script type="text/javascript">
-    // Submit function to be called, after reCAPTCHA was successful.
-    var onSubmit_{{ widget_uuid }} = function(token) {
-        console.log("reCAPTCHA validated for 'data-widget-uuid=\"{{ widget_uuid }}\"'");
-        $('input[name="g-recaptcha-response"]').val(token);
-    };
-</script>
