--- lametro/templates/captcha/widget_v2_checkbox.html
@@ -1,6 +1,0 @@
-{% include "captcha/includes/js_v2_checkbox.html" %}
-<div class="g-recaptcha"
-    {% for name, value in widget.attrs.items %}{% if value is not False %} {{ name }}{% if value is not True %}="{{ value|stringformat:'s' }}"{% endif %}{% endif %}{% endfor %}
->
-</div>
-<input type="hidden" name="g-recaptcha-response" />
